fix: rename folder, formatTime util
