+++ packages/contact-center/cc-components/src/components/UserState/user-state.scss
