+++ packages/contact-center/cc-components/src/index.ts
@@ -1,5 +1,5 @@
-import UserStateComponent from './components/user-state/user-state';
-import {IUserState} from './components/user-state/user-state.types';
+import UserStateComponent from './components/UserState/user-state';
+import {IUserState} from './components/UserState/user-state.types';
 import '@momentum-ui/icons/css/momentum-ui-icons.min.css';
 import '@momentum-ui/core/css/momentum-ui.min.css';
 import '@momentum-ui/web-components';
