+++ packages/contact-center/cc-components/src/wc.ts
@@ -1,5 +1,5 @@
 import r2wc from '@r2wc/react-to-web-component';
-import UserStateComponent from './components/user-state/user-state';
+import UserStateComponent from './components/UserState/user-state';
 import '@momentum-ui/web-components';
 import '@momentum-ui/icons/css/momentum-ui-icons.min.css';
 import '@momentum-ui/core/css/momentum-ui.min.css';
